Refactor physics pipeline and SVG serializers
Extracted physics overflow eviction and quiescence tracking into dedicated modules, simplifying PhysicsPipelineStore while preserving behavior. Split ordinal scene-node SVG rendering out of SceneToSVG into SceneToSVGOrdinal and moved shared SVG helpers into sceneToSVGShared to keep SSR alignment and file-size limits clean. Also moved physics builder tests into their own suite and applied small follow-up fixes/documentation updates (mode wording, rerun config memoization, keyboard tab handling, and regenerated size benchmarks).
