+++ src/components/charts/shared/types.ts
@@ -47,7 +47,7 @@ export type LinkedBrushProp =
   | { name: string; xField?: string; yField?: string }
 
 /**
- * Chart display mode — controls default chrome, size, and interaction.
+ * Chart display mode — controls default decoration, size, and interaction.
  * User-provided props always override mode defaults.
  */
 export type ChartMode = "primary" | "context" | "sparkline" | "mobile"
@@ -115,7 +115,7 @@ export interface ResolvedMobileInteractionConfig {
  * Base props shared across all chart components
  */
 export interface BaseChartProps {
-  /** Display mode: "primary" (full chrome), "context" (compact), "sparkline" (inline), or "mobile" (touch-first) */
+  /** Display mode: "primary" (full decoration), "context" (compact), "sparkline" (inline), or "mobile" (touch-first) */
   mode?: ChartMode
   /** Chart width in pixels. Default: 600 */
   width?: number
